requirements.txt
alembic==1.13.1
azure-core==1.30.1
azure-storage-blob==12.20.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
mysql-connector==2.2.9
mysql-connector-python==8.3.0
mysql-connector-python-rf==2.2.2
mysqlclient==2.2.4
pycparser==2.22
PyMySQL==1.1.0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
WTForms==3.1.2
